datapath: move IPIP device setup out of loader
Move IPIP device setup and the shared netlink helpers into the Linux
device package. Invoke device setup directly from the datapath
orchestrator. Re-run setup when reconciliation is triggered to handle MTU
updates and device recreation while preserving retry behavior.

This keeps the loader focused on BPF loading while the orchestrator owns
device setup ordering.
