+++ pkg/datapath/loader/netlink_test.go
@@ -53,63 +53,6 @@ func mustXDPProgram(t *testing.T, name string) *ebpf.Program {
 	return p
 }
 
-func TestPrivilegedSetupDev(t *testing.T) {
-	testutils.PrivilegedTest(t)
-	logger := hivetest.Logger(t)
-
-	sysctl := sysctl.NewDirectSysctl(afero.NewOsFs(), "/proc")
-
-	prevConfigEnableIPv4 := option.Config.EnableIPv4
-	prevConfigEnableIPv6 := option.Config.EnableIPv6
-	t.Cleanup(func() {
-		option.Config.EnableIPv4 = prevConfigEnableIPv4
-		option.Config.EnableIPv6 = prevConfigEnableIPv6
-	})
-	option.Config.EnableIPv4 = true
-	option.Config.EnableIPv6 = true
-
-	ns := netns.NewNetNS(t)
-
-	ns.Do(func() error {
-		ifName := "dummy"
-		dummy := &netlink.Dummy{
-			LinkAttrs: netlink.LinkAttrs{
-				Name: ifName,
-			},
-		}
-		err := netlink.LinkAdd(dummy)
-		require.NoError(t, err)
-
-		err = enableForwarding(logger, sysctl, dummy)
-		require.NoError(t, err)
-
-		enabledSettings := [][]string{
-			{"net", "ipv6", "conf", ifName, "forwarding"},
-			{"net", "ipv4", "conf", ifName, "forwarding"},
-			{"net", "ipv4", "conf", ifName, "accept_local"},
-		}
-		disabledSettings := [][]string{
-			{"net", "ipv4", "conf", ifName, "rp_filter"},
-			{"net", "ipv4", "conf", ifName, "send_redirects"},
-		}
-		for _, setting := range enabledSettings {
-			s, err := sysctl.Read(setting)
-			require.NoError(t, err)
-			require.Equal(t, "1", s)
-		}
-		for _, setting := range disabledSettings {
-			s, err := sysctl.Read(setting)
-			require.NoError(t, err)
-			require.Equal(t, "0", s)
-		}
-
-		err = netlink.LinkDel(dummy)
-		require.NoError(t, err)
-
-		return nil
-	})
-}
-
 // TestPrivilegedSetupBaseDeviceIPv6NotTentative checks that the IPv6 link-local
 // addresses of the base devices are usable as soon as setupBaseDevice returns.
 // Bringing a link up before turning ARP off leaves the kernel-generated
@@ -513,59 +456,6 @@ func TestPrivilegedAddHostDeviceAddr(t *testing.T) {
 
 		err = netlink.LinkDel(dummy)
 		require.NoError(t, err)
-
-		return nil
-	})
-}
-
-func TestPrivilegedSetupIPIPDevices(t *testing.T) {
-	testutils.PrivilegedTest(t)
-
-	logger := hivetest.Logger(t)
-
-	sysctl := sysctl.NewDirectSysctl(afero.NewOsFs(), "/proc")
-
-	ns := netns.NewNetNS(t)
-	ns.Do(func() error {
-		err := setupIPIPDevices(logger, sysctl, true, true, 1500)
-		require.NoError(t, err)
-
-		dev4, err := safenetlink.LinkByName(defaults.IPIPv4Device)
-		require.NoError(t, err)
-		require.Equal(t, 1480, dev4.Attrs().MTU)
-
-		dev6, err := safenetlink.LinkByName(defaults.IPIPv6Device)
-		require.NoError(t, err)
-		require.Equal(t, 1452, dev6.Attrs().MTU)
-
-		err = setupIPIPDevices(logger, sysctl, false, false, 1500)
-		require.NoError(t, err)
-
-		_, err = safenetlink.LinkByName(defaults.IPIPv4Device)
-		require.Error(t, err)
-
-		_, err = safenetlink.LinkByName(defaults.IPIPv6Device)
-		require.Error(t, err)
-
-		err = setupIPIPDevices(logger, sysctl, true, true, 1480)
-		require.NoError(t, err)
-
-		dev4, err = safenetlink.LinkByName(defaults.IPIPv4Device)
-		require.NoError(t, err)
-		require.Equal(t, 1460, dev4.Attrs().MTU)
-
-		dev6, err = safenetlink.LinkByName(defaults.IPIPv6Device)
-		require.NoError(t, err)
-		require.Equal(t, 1432, dev6.Attrs().MTU)
-
-		err = setupIPIPDevices(logger, sysctl, false, false, 1480)
-		require.NoError(t, err)
-
-		_, err = safenetlink.LinkByName(defaults.IPIPv4Device)
-		require.Error(t, err)
-
-		_, err = safenetlink.LinkByName(defaults.IPIPv6Device)
-		require.Error(t, err)
 
 		return nil
 	})
